+++ QuantumSim/sim/kernels/kernelsAVX.h
@@ -44,7 +44,7 @@ using namespace std;
 
 constexpr float kH = 0.707106781;
 
-using idx_size = size_t;
+using idx_size = unsigned long long;
 
 constexpr cmplx kTGate[8] = {1, {kH, kH}, {0,1}, {-kH, kH}, -1, {-kH, -kH}, {0, -1}, {kH, -kH}};
 constexpr float kTGate_re[8] = {1, kH, 0, -kH, -1, -kH, 0, kH};
